backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
SQLAlchemy
alembic
python-jose
passlib[bcrypt]
pydantic
email-validator
pydantic[email]
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
scikit-learn
joblib
python-multipart

